q-news/requirements.txt
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
akshare>=1.16
feedparser>=6.0
anthropic>=0.40
openai>=1.50
python-dotenv>=1.0
